Split templates in their own files.
As a side effet this fixes build errors on Ruby 1.9.1 when LANG=C.
DhMakeRuby was calling File.read() on its own file to get the templates
data, but when LANG=C this makes Ruby 1.9 crash because the file is
UTF-8. Since the templates themselves are us-ascii, putting them in
their own files avoids this problem.