dh-make-ruby: avoid tests depending on @builddeps@
@builddeps@ will contain gem2deb ~100% of the time, and we don't want to
pull in that many dependencies when running the tests (this is the
reason gem2deb-test-runner was split in its own package in the first
place).

To alleviate the situation and try to get tests that work out of the box
most of the time, make gem2deb-test-runner depend on packages usually
used for test suites: ruby-rspec, ruby-test-unit, and rake.