Remove assignments to Encoding.default_external
This removes warnings that show up when running tests. According to the
Ruby documentation, setting Encoding.default_external after a program in
has started can leave strings created before the assignment in an
invalid state (i.e. in a previous default encoding).

This requires the encoding to be set to UTF-8 in the shebang, and LANG
to be set to C.UTF-8 in Rakefile.

GBp-Dch: ignore